major update, added config file
I finished this a while ago but I forgot to add it, sorry!

+++ on idle start archiveteam and torrent client.py
@@ -1,40 +1,63 @@
 import os
 import win32api
+import configparser
 from time import sleep
+import subprocess
 
-# important! 
-# replace the below text where shown with the actual names/paths for the files 
-# replace \ with \\ or python will get confused
-# the replaced text must be in quotes "like this", see example
-wgcf = "REPLACE THIS TEXT WITH WIREGUARD CONFIG PATH" # example: "C:\\Program Files\\WireGuard\\Data\\Configurations\\wgcf-profile.conf.dpapi"
-atvm = "REPLACE THIS TEXT WITH THE NAME OF THE ARCHIVETEAM VM" # example: "archiveteam-warrior-3.2"
+config = configparser.ConfigParser(allow_no_value=True)
+
+# checks if config file exists, if it doesn't, create new one with booleanss, if it does, read it
+if not os.path.isfile('cfg.ini'):
+    # defauits
+    config['booleans'] = {'# change to 1 for on, 0 for off':None, 'wireguard':'1', 'wireguard turns off after idle ends':'1', 'qbittorrent':'1', 'qbittorrent turns off after idle ends':'1', 'archiveteam VM':'1', 'archiveteam VM turns off after idle ends':'1', 'ytsync':'1'}
+    config['paths'] = {'# you do not need to fill in ones that are turned off in the boolean section!':None, 'wireguard executable path':r'C:\Program Files\WireGuard\wireguard.exe', 'wireguard config path':r'C:\Program Files\WireGuard\Data\Configurations\wgcf-profile.conf.dpapi', 'qbittorrent executable path':r'C:\Program Files\qBittorrent\qbittorrent.exe', 'virtualbox executable path':r'C:\Program Files\Oracle\VirtualBox\VBoxManage.exe', 'ytsync path':r'G:\music\playlist sync.lnk'}
+    config['misc'] = {'archiveteam warrior vm name':'archiveteam-warrior-3.2', '# the duration should be in seconds':None, 'time before idle':'1200'}
+    # creating file requires admin, reading might too
+    config.write(open('cfg.ini', 'w'))
+else: config.read('cfg.ini')
 
 while(True): # this is so that the idle checking still continues if the computer becomes temporarily active
-
     fixedposition = win32api.GetCursorPos() # a lot of this type of code could probably be put into a dedicated function but I can't be bothered
     mousetimer = 0
-
-    while(mousetimer < 600): #default is 600 seconds (10 minutes), can be changed to your preference
+    # increases mousetimer until it reaches a set value, mouse movement resets it to 0
+    while(mousetimer < int(config['misc']['time before idle'])):
         position = win32api.GetCursorPos()
         if fixedposition != position:
             mousetimer = 0
             fixedposition = win32api.GetCursorPos()
-        sleep(0.05)
         mousetimer = round((mousetimer + 0.05), 2) # rounding is to remove floating point errors, not very important but it looks nice
+        sleep(0.05)
 
-    # runs the programs, REQUIRES ADMIN, assumes the paths are the default which may change in future updates
-    os.popen(f'"C:\\Program Files\\WireGuard\\wireguard.exe" /installtunnelservice "{wgcf}"')
-    os.popen('"C:\\Program Files\\qBittorrent\\qbittorrent.exe"')
-    os.popen(f'"C:\\Program Files\\Oracle\\VirtualBox\\VBoxManage.exe" startvm "{atvm}"')
+    # runs the programs, REQUIRES ADMIN
+    # these should be done without the repeated if statements but I (mistakenly) thought it wouldn't matter
+    # subprocess doesn't like working in the background so it is commented out
+    if int(config['booleans']['wireguard']) == 1:
+        os.popen(fr'"{config['paths']['wireguard executable path']}" /installtunnelservice "{config['paths']['wireguard config path']}"')
+        # subprocess.Popen(fr'"{config['paths']['wireguard executable path']}" /installtunnelservice "{config['paths']['wireguard config path']}"',shell=False, stdin=None, stdout=None, stderr=None)
+    if int(config['booleans']['qbittorrent']) == 1:
+        os.popen(fr'"{config['paths']['qbittorrent executable path']}"')
+        # subprocess.Popen(fr'"{config['paths']['qbittorrent executable path']}"',shell=False, stdin=None, stdout=None, stderr=None)
+    if int(config['booleans']['archiveteam VM']) == 1:
+        os.popen(fr'"{config['paths']['virtualbox executable path']}" startvm "{config['misc']['archiveteam warrior vm name']}"')
+        # subprocess.Popen(fr'"{config['paths']['virtualbox executable path']}" startvm "{config['misc']['archiveteam warrior vm name']}"',shell=False, stdin=None, stdout=None, stderr=None)
+    if int(config['booleans']['ytsync']) == 1:
+        os.popen(fr'"{config['paths']['ytsync path']}"')
+        # subprocess.Popen(fr'"{config['paths']['ytsync path']}"',shell=True, stdin=None, stdout=None, stderr=None)
 
     # for closing programs when coming out of idle, shamefully stolen from stackoverflow
+    # also could be done without the repeated if statements
     fixedposition = win32api.GetCursorPos()
     while(True):
         position = win32api.GetCursorPos()
         if fixedposition != position: # this code also requires admin unfortunately
-            os.popen(f'"C:\\Program Files\\Oracle\\VirtualBox\\VBoxManage.exe" controlvm "{atvm}" acpipowerbutton')
-            os.popen('taskkill /im "qbittorrent.exe"')
-            # comment the below if you don't want it to turn off the vpn when computer becomes active
-            os.popen(f'"C:\\Program Files\\WireGuard\\wireguard.exe" /uninstalltunnelservice "{atvm}"')
+            if int(config['booleans']['archiveteam VM turns off after idle ends']) == 1:
+                os.popen(fr'"{config['paths']['virtualbox executable path']}" controlvm "{config['misc']['archiveteam warrior vm name']}" acpipowerbutton')
+                # subprocess.Popen(fr'"{config['paths']['virtualbox executable path']}" controlvm "{config['misc']['archiveteam warrior vm name']}" acpipowerbutton',shell=False, stdin=None, stdout = None, stderr=None)
+            if int(config['booleans']['qbittorrent turns off after idle ends']) == 1:
+                os.popen('taskkill /im "qbittorrent.exe"')
+                # subprocess.Popen('taskkill /im "qbittorrent.exe"',shell=False, stdin=None, stdout=None, stderr=None)
+            if int(config['booleans']['wireguard turns off after idle ends']) == 1:
+                os.popen(fr'"{config['paths']['wireguard executable path']}" /uninstalltunnelservice "{config['paths']['wireguard config path']}"')
+                # subprocess.Popen(fr'"{config['paths']['wireguard executable path']}" /uninstalltunnelservice "{config['paths']['wireguard config path']}"',shell=False, stdin=None, stdout=None, stderr=None)
             break
         sleep(0.05)
